requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
python-jose
pydantic-settings